+++ vera/__init__.py
@@ -1,4 +1,4 @@
 """Vera: a programming language designed for LLMs."""
 
-__version__ = "0.0.154"
-version = "0.0.154"
+__version__ = "0.0.155"
+version = "0.0.155"